- update minimum SFML version to 2.4 (fix deprecated get/setColor) - embed sfeMovie - update cmake to add FFmpeg FindFFMPEG.cmake - update deprecated code for FFMPEG3 upgrade in sfeMovie (90%) - remove unused code from sfeMovie (SubtitleStream)
- add isUsable to item
- add touch support (untested)
- add UIText interface
- execute action given drawable id and action name
- refactor buttons
- refactor cel decoding (fix small font)
- refactor input events code
- refactor setText actions into one (text.setText)
- refactor StringButton, Text2
- update android minimum version
- update itemClass description to support skipping matches
- update UIObject interface to add method to get action

- update gamefiles
- resize level panel if user panels are open/closed and if size = 640x480
- update item prefixes and descriptions
- update level right click

+++ src/Game/ItemLocation.h
@@ -23,7 +23,7 @@ public:
 	ItemCoordInventory(const std::string& playerId_, size_t inventoryIdx_, size_t itemIdx_)
 		: playerId(playerId_),
 		inventoryIdx(((int16_t)inventoryIdx_) & 0x7FFF),
-		itemIdx(itemIdx_) {}
+		itemIdx((int16_t)itemIdx_) {}
 
 	ItemCoordInventory(const std::string& playerId_, size_t inventoryIdx_, const ItemXY& itemXY_)
 		: playerId(playerId_),
